0. Print list

+++ 0x12-singly_linked_lists/lists.h
@@ -5,17 +5,18 @@
 #include <stdlib.h>
 #include <string.h>
 
-
 /**
  * struct list_s - singly linked list
- * @str: malloc string
- * @length: string length
- * @nxt: pointer to the next sring
+ * @str: string - (malloc'ed string)
+ * @len: length of the string
+ * @next: points to the next node
+ *
+ * Description: singly linked list node structure
  */
 typedef struct list_s
 {
 	char *str;
-	size_t len;
+	unsigned int len;
 	struct list_s *next;
 } list_t;
